+++ CHANGELOG.md
@@ -42,6 +42,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Reinserting a breakpoint after the last breakpoint on a page was removed now
+  refreshes the retained shadow page and maps it back into the target view.
+
 ## [0.7.0] - 2026-06-24
 
 ### Changed